Added location services

+++ static/js/components/NodeSearch.js
@@ -23,11 +23,14 @@ let nodeSearch = Vue.component('node-search', {
             }, 120);
         },
         deleteSearch: function () {
-            if (this.id > 0)
-                this.$parent.deleteNode(this.id);
-            else
-                this.nodeInput = '';
-            removeMarker(this.id);
+            if (this.nodeInput !== '') {
+                if (this.id > 0)
+                    this.$parent.deleteNode(this.id);
+                else
+                    this.nodeInput = '';
+
+                removeMarker(this.id);
+            }
         }
     },
     props: {id: Number, index: Number},
